🚚 api: move get certification-report-serializer to src directory
