+++ api/tests/certification/shared/unit/infrastructure/serializers/jsonapi/certification-report-serializer_test.js
@@ -1,9 +1,9 @@
-import { expect, domainBuilder } from '../../../../test-helper.js';
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/certification-report-serializer.js';
+import { expect, domainBuilder } from '../../../../../../test-helper.js';
+import * as serializer from '../../../../../../../src/certification/shared/infrastructure/serializers/jsonapi/certification-report-serializer.js';
 import {
   CertificationIssueReportCategory,
   CertificationIssueReportSubcategories,
-} from '../../../../../src/certification/shared/domain/models/CertificationIssueReportCategory.js';
+} from '../../../../../../../src/certification/shared/domain/models/CertificationIssueReportCategory.js';
 
 describe('Unit | Serializer | JSONAPI | certification-report-serializer', function () {
   describe('#serialize()', function () {
